Refactoring and inplementation of keyset features started

+++ eventrecorderserver/eventrecorder/src/main/java/com/epam/eventrecorder/user/service/DeleteUserService.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.epam.eventrecorder.user.dao.UserDao;
+import com.epam.eventrecorder.user.domain.User;
 
 @Service
 public class DeleteUserService {
@@ -12,6 +13,12 @@ public class DeleteUserService {
     private UserDao userDao;
 
     public void deleteUser(Long userId) {
-        userDao.deleteUser(userId);
+        User user = userDao.findUserById(userId);
+        if (user.getExperiments().isEmpty() && user.getKeysets().isEmpty()) {
+            userDao.deleteUser(userId);
+        } else {
+            throw new IllegalArgumentException("User " + user + " can not be deleted. Delete experiments and keysets before.");
+        }
+
     }
 }
